fix(code highlight): update prismjs.js
