+++ scripts/helper.js
@@ -12,10 +12,6 @@ hexo.extend.helper.register('customToc', function(){
    return require("./lib/custom_toc.js").call(hexo,...arguments);
  });
 
-hexo.extend.helper.register('prism', function(){
-  return require("./lib/prism/index.js").call(hexo,...arguments);
-});
-
 hexo.extend.helper.register('jsLsload', function(){
   return require("./lib/ls/js_lsload.js").call(hexo,...arguments);
 });
