+++ source/css/_files/plus/post.css
@@ -144,7 +144,6 @@ body.mdui-drawer-body-left .post-bg-color {
 }
 
 #post-content.mdui-typo pre code {
-  display: block;
   padding: 12px;
   overflow: inherit;
 }
